==> 3_Evaluate.ipynb <==
import numpy as np
import pandas as pd

from params import *

results_NULL = pd.read_csv('results/results_online_NULL.txt', header = None)
results_OPTIMAL = pd.read_csv('results/results_online_OPTIMAL.txt', header = None)
results_SEP = pd.read_csv('results/results_online_SEP.txt', header = None)
results_SHR = pd.read_csv('results/results_online_SHR.txt', header = None)
results_SHR_CLUSTER = pd.read_csv('results/results_online_SHR_CLUSTER.txt', header = None)
results_SHR_ROBUST = pd.read_csv('results/results_online_SHR_ROBUST.txt', header = None)
results_SHR_FULL = pd.read_csv('results/results_online_SHR_FULL.txt', header = None)

# compare the objective functions

for passenger_percentage in [0, 0.2, 0.4, 0.6, 0.8, 1]:
    print(f'SEP (passenger percentage {passenger_percentage})', np.mean(results_NULL[2] + results_SEP[results_SEP[2]==passenger_percentage][3]))
    
print('SHR', np.mean(results_NULL[2] + results_SHR[2]))
print('SHR-CLUSTER', np.mean(results_NULL[2] + results_SHR_CLUSTER[2]))
print('SHR-ROBUST', np.mean(results_NULL[2] + results_SHR_ROBUST[2]))
print('SHR-FULL', np.mean(results_NULL[2] + results_SHR_FULL[2]))

# compare the decision error

print('SHR', np.mean(1-(results_NULL[2]+results_SHR[2])/results_OPTIMAL[2]))
print('SHR-CLUSTER', np.mean(1-(results_NULL[2]+results_SHR_CLUSTER[2])/results_OPTIMAL[2]))
print('SHR-ROBUST', np.mean(1-(results_NULL[2]+results_SHR_ROBUST[2])/results_OPTIMAL[2]))
print('SHR-FULL', np.mean(1-(results_NULL[2]+results_SHR_FULL[2])/results_OPTIMAL[2]))

# compare the coverage percentage

print('SHR', np.mean(np.minimum(1, results_SHR[3] / N_incoming)))
print('SHR-CLUSTER', np.mean(np.minimum(1, results_SHR_CLUSTER[3] / N_incoming)))
print('SHR-ROBUST', np.mean(np.minimum(1, results_SHR_ROBUST[3] / N_incoming)))
print('SHR-FULL', np.mean(np.minimum(1, results_SHR_FULL[3] / N_incoming)))
